Split the import queue in three files
